data-pipeline-2/requirements.txt
pandas
pymysql
python-dotenv
boto3
pyodbc
freezegun
pytest